test(subagent): split the native-leaf tests so the windows lane compiles
